\name{RcppGsl-package}
\alias{RcppGsl-package}
\alias{RcppGsl}
\docType{package}
\title{
glue between Rcpp and gsl
}
\description{
glue between Rcpp and gsl
}
\details{
\tabular{ll}{
Package: \tab RcppGsl\cr
Type: \tab Package\cr
Version: \tab 0.0\cr
Date: \tab 2010-04-04\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
 Romain Francois and Dirk Eddelbuettel

Maintainer: Romain and Dirk <RomainAndDirk@r-enthusiasts.com>
}
\references{
	GSL: Gnu scientific library: \url{http://www.gnu.org/software/gsl/}
}
\keyword{ package }

