\name{RcppGSL-package}
\alias{RcppGSL-package}
\alias{RcppGSL}
\docType{package}
\title{
  Glue between Rcpp and the GNU GSL
}
\description{
  Glue between Rcpp and the GNU GSL
}
\details{
  \tabular{ll}{
	Package: \tab RcppGSL\cr
	Type: \tab Package\cr
	Version: \tab 0.0\cr
	Date: \tab 2010-04-04\cr
	License: \tab GPL (>= 2)\cr
	LazyLoad: \tab yes\cr
  }
}
\author{
  Romain Francois and Dirk Eddelbuettel

  Maintainer: Dirk Eddelbuettel and Romain Francois <RomainAndDirk@r-enthusiasts.com>
}
\references{
  GSL: GNU Scientific Library: \url{http://www.gnu.org/software/gsl/}
}
\keyword{ package }

