\name{RcppGSL-package}
\alias{RcppGSL-package}
\alias{RcppGSL}
\docType{package}
\title{\packageTitle{RcppGSL}}
\description{\packageDescription{RcppGSL}}
\author{\packageAuthor{RcppGSL}}
\section{Maintainer}{\packageMaintainer{RcppGSL}}
\references{GSL: GNU Scientific Library: \url{http://www.gnu.org/software/gsl/}}
\keyword{ package }

